+++ packages/mongodb/plugin.ts
@@ -3,7 +3,7 @@ import { setupProvider } from '@zemble/core/utils/setupProvider'
 import Routes from '@zemble/routes'
 import { MongoClient } from 'mongodb'
 
-import type { MongoClientOptions } from 'mongodb'
+import type { Db, MongoClientOptions } from 'mongodb'
 
 interface MongodbClientConfig extends Zemble.GlobalConfig {
   readonly url?: string
@@ -28,7 +28,10 @@ declare global {
 
     interface Providers {
       // eslint-disable-next-line functional/prefer-readonly-type
-      'mongodb': MongoClient | undefined
+      'mongodb': {
+        readonly client: MongoClient
+        readonly db: Db
+      } | undefined
     }
   }
 }
@@ -37,34 +40,52 @@ const defaultConfig = {
   url: process.env.MONGO_URL ?? 'mongodb://localhost:27017',
 } satisfies MongodbClientConfig
 
+const regexToHidePassword = /(?<=mongodb\+srv:\/\/[^:]+:)[^@]+/
+
 export default new PluginWithMiddleware<MongodbClientConfig, typeof defaultConfig>(
   __dirname,
   async ({
-    app, config, plugins,
+    app, config, plugins, context,
   }) => {
     if (process.env.DEBUG) {
-      const regex = /(?<=mongodb\+srv:\/\/[^:]+:)[^@]+/
-      console.log('Connecting to MongoDB', config.url.replace(regex, '***'))
+      context.logger.log('Connecting to MongoDB', config.url.replace(regexToHidePassword, '***'))
     }
 
     // we create a global mongodb client for the app, which is also used for all plugins that don't have a custom config
-    const client = await MongoClient.connect(config.url, config.options)
+    const defaultClient = new MongoClient(config.url, config.options)
 
-    client.on('error', (error) => {
-      console.error(error, 'MongoDB error')
+    defaultClient.on('error', (error) => {
+      context.logger.error('MongoDB error', error)
     })
 
-    if (process.env.DEBUG) {
-      console.log('Connected to MongoDB!')
-    }
+    context.logger.log('Connected to MongoDB')
+
+    await defaultClient.connect()
 
     await setupProvider({
       app,
       initializeProvider: async (customConfig) => {
         const pluginCustomConfig = customConfig?.config
-        return pluginCustomConfig
-          ? MongoClient.connect(pluginCustomConfig.url, pluginCustomConfig.options)
-          : client
+
+        if (pluginCustomConfig) {
+          const customClient = new MongoClient(config.url, config.options)
+
+          customClient.on('error', (error) => {
+            context.logger.error('MongoDB error', error)
+          })
+
+          await customClient.connect()
+
+          return {
+            client: customClient,
+            db: customClient.db(),
+          }
+        }
+
+        return {
+          client: defaultClient,
+          db: defaultClient.db(),
+        }
       },
       providerKey: 'mongodb',
       plugins,