+++ packages/bull/CHANGELOG.md
@@ -1,5 +1,11 @@
 # zemble-plugin-bull
 
+## 0.0.24
+
+### Patch Changes
+
+- 6479619: Fix serialization for bull, use global db instance for mongodb
+
 ## 0.0.23
 
 ### Patch Changes