+++ packages/cms-users/CHANGELOG.md
@@ -1,5 +1,13 @@
 # zemble-plugin-cms-users
 
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies [6479619]
+  - @zemble/mongodb@0.0.9
+  - zemble-plugin-cms@0.0.24
+
 ## 0.0.23
 
 ### Patch Changes