+++ packages/ignite/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @zemble/ignite
 
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies [6479619]
+  - zemble-plugin-bull@0.0.24
+
 ## 0.0.15
 
 ### Patch Changes